+++ vscode/src/chat/ChatViewProvider.ts
@@ -76,14 +76,12 @@ export type Config = Pick<
  */
 const SAFETY_PROMPT_TOKENS = 100
 
-export interface ChatViewProviderWebview extends Omit<vscode.Webview, 'postMessage'> {
-    postMessage(message: ExtensionMessage): Thenable<boolean>
-}
-
 export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disposable, IdleRecipeRunner {
     private isMessageInProgress = false
     private cancelCompletionCallback: (() => void) | null = null
-    public webview?: ChatViewProviderWebview
+    public webview?: Omit<vscode.Webview, 'postMessage'> & {
+        postMessage(message: ExtensionMessage): Thenable<boolean>
+    }
 
     private currentChatID = ''
     private inputHistory: string[] = []
@@ -331,6 +329,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
     private sendPrompt(promptMessages: Message[], responsePrefix = ''): void {
         this.cancelCompletion()
         void vscode.commands.executeCommand('setContext', 'cody.reply.pending', true)
+        this.editor.controllers.inline.setResponsePending(true)
 
         let text = ''
 
@@ -339,6 +338,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                 text += content
                 const displayText = reformatBotMessage(text, responsePrefix)
                 this.transcript.addAssistantResponse(displayText)
+                this.editor.controllers.inline.reply(displayText, 'streaming')
                 this.sendTranscript()
                 return Promise.resolve()
             },
@@ -361,6 +361,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                     // TODO(keegancsmith) guardrails may be slow, we need to make this async update the interaction.
                     highlightedDisplayText = await this.guardrailsAnnotateAttributions(highlightedDisplayText)
                     this.transcript.addAssistantResponse(text || '', highlightedDisplayText)
+                    this.editor.controllers.inline.reply(highlightedDisplayText, 'complete')
                 }
                 void this.onCompletionEnd()
             },
@@ -407,6 +408,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                 // We ignore embeddings errors in this instance because we're already showing an
                 // error message and don't want to overwhelm the user.
                 this.onCompletionEnd(true)
+                void this.editor.controllers.inline.error()
                 console.error(`Completion request failed: ${err}`)
             },
         })
@@ -424,6 +426,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         void this.saveTranscriptToChatHistory()
         this.sendChatHistory()
         void vscode.commands.executeCommand('setContext', 'cody.reply.pending', false)
+        this.editor.controllers.inline.setResponsePending(false)
         if (!ignoreEmbeddingsError) {
             this.logEmbeddingsSearchErrors()
         }
